backend/requirements.txt
accelerate 
annotated-types 
asgiref 
bitsandbytes 
CacheControl 
cachetools 
certifi 
cffi 
channels
charset-normalizer 
colorama 
cryptography 
Django 
django-cors-headers 
djangorestframework 
djangorestframework-simplejwt 
dotenv 
filelock 
firebase-admin 
fsspec 
google-ai-generativelanguage 
google-api-core 
google-api-python-client 
google-auth 
google-auth-httplib2 
google-auth-oauthlib 
google-cloud-core 
google-cloud-firestore 
google-cloud-storage 
google-crc32c 
google-generativeai 
google-resumable-media 
googleapis-common-protos 
grpcio 
grpcio-status 
gunicorn 
httplib2 
huggingface-hub 
idna 
Jinja2 
MarkupSafe 
mpmath 
msgpack 
networkx 
numpy 
oauthlib 
packaging 
pillow 
proto-plus 
protobuf 
psutil 
psycopg2-binary 
pyasn1 
pyasn1_modules 
pycparser 
pydantic 
pydantic_core 
PyJWT 
pyparsing 
python-dotenv 
PyYAML 
regex 
requests 
requests-oauthlib 
resend 
rsa 
safetensors 
setuptools 
sqlparse 
sympy 
tokenizers 
torch 
torchaudio 
torchvision 
tqdm 
transformers 
typing_extensions 
tzdata 
uritemplate 
urllib3 
unsloth
uvicorn[standard]